# server/swagger.yml
openapi: 3.0.0
info:
  description: CUET Transport Automation System API
  version: 1.0.0
  title: CUET Transport Automation System API
  contact:
    email: tajircuet@gmail.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: /ctas/v1
tags:
  - name: users
    description: Users related api.
  - name: buses
    description: Buses related api.
  - name: drivers
    description: Drivers related api.
  - name: routes
    description: Routes related api
paths:
  /buses:
    get:
      tags:
        - buses
      summary: Get Bus list
      description: Returns a list of buses
      operationId: getBuses
      security:
        - AccessTokenAuth: []
      responses:
        200:
          description: OK
          content: 
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    format: int32
                  data:
                    type: object
                    properties:
                      items:
                        type: array
                        items:
                          $ref: '#/components/schemas/Bus'
    post:
      tags:
        - buses
      summary: Add new bus
      description: Add new bus
      operationId: addBus
      parameters:
        - name: name
          in: query
          required: true
          description: Bus name
          type: string
        - name: status
          in: query
          required: true
          description: Bus Condition good
          type: boolean
    put:
      tags:
        - buses
      summary: Updates bus information
      description: Updates bus information
      operationId: updateBus
      parameters:
        - name: id
          in: query
          required: true
          description: Bus id
          type: integer
        - name: name
          in: query
          required: true
          description: Bus name
          type: string
        - name: status
          in: query
          required: true
          description: Bus Condition good
          type: boolean
  /drivers:
    get:
      tags:
        - drivers
      summary: Get Drivers list
      description: Returns a list of drivers
      operationId: getDrivers
      security:
        - AccessTokenAuth: []
      responses:
        200:
          description: OK
          content: 
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    format: int32
                  data:
                    type: object
                    properties:
                      items:
                        type: array
                        items:
                          $ref: '#/components/schemas/Driver'
    post:
      tags:
        - driver
      summary: Add new driver
      description: Add new driver
      operationId: addDriver
      parameters:
        - name: name
          in: query
          required: true
          description: Driver name
          type: string
        - name: address
          in: query
          required: true
          description: Driver address
          type: string
        - name: licenseStatus
          in: query
          required: true
          description: Driver License Status
          type: boolean
    put:
      tags:
        - driver
      summary: Updates driver information
      description: Updates driver information
      operationId: updateDriver
      parameters:
        - name: id
          in: query
          required: true
          description: Driver id
          type: integer
        - name: name
          in: query
          required: true
          description: Driver name
          type: string
        - name: address
          in: query
          required: true
          description: Driver address
          type: string
        - name: licenseStatus
          in: query
          required: true
          description: Driver License Status
          type: boolean
  /bus-routes:
    get:
      tags:
        - routes
      summary: Get Routes list
      description: Returns a list of available routes
      operationId: getRoutes
      security:
        - AccessTokenAuth: []
      responses:
        200:
          description: OK
          content: 
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    format: int32
                  data:
                    type: object
                    properties:
                      items:
                        type: array
                        items:
                          $ref: '#/components/schemas/Route'
  /users/login:
    post:
      tags:
        - users
      summary: Logs user into the system
      description: Use username and password to login.
      operationId: login
      parameters:
        - name: username
          in: query
          required: true
          description: The user name for login.
          schema:
            $ref: '#/components/schemas/User/properties/username'
        - name: password
          in: query
          required: true
          description: The password for login in clear text.
          schema:
            $ref: '#/components/schemas/User/properties/password'
      responses:
        200:
          description: OK
          headers:
            X-Rate-Limit:
              schema:
                type: integer
                format: int32
              description: calls per hour allowed by the user.
            X-Expires-After:
              schema:
                type: string
                format: date-time
              description: date in UTC when token expires.
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    format: int32
                  data:
                    type: object
                    properties:
                      accessToken:
                        type: string
                        example: admin-token
        400:
          $ref: '#/components/responses/InvalidUser'
  /users/logout:
    post:
      tags:
        - users
      summary: Logs out current logged in user session
      description: Log out current session.
      operationId: logout
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    format: int32
  /users/info:
    get:
      tags:
        - users
      summary: Get Users information
      description: Returns user information
      operationId: getUserInfo
      security:
        - AccessTokenAuth: []
      responses:
        200:
          description: OK
          content: 
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      id:
                        type: integer
                        format: int64
                      username:
                        type: string
                      role:
                        type: integer
                        format: int64
  /schedule:
    get:
      tags:
        - bus
      summary: Get Bus Schedule
      description: Returns bus schedule
      operationId: getSchedule
      security:
        - AccessTokenAuth: []
      responses:
        200:
          description: OK
          content: 
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      schedule:
                        type: array
                        items:
                          id:
                            type: integer
                          bus:
                            type: integer
                          driver:
                            type: integer
                          route:
                            type: integer
                          datetime:
                            type: string
                          isStandBy:
                            type: boolean
    post:
      tags:
        - bus
      summary: Add new schedule for a bus
      description: Inserts new schedule record for a bus
      operationId: addSchedule
      parameters:
        - name: bus
          in: query
          required: true
          description: Bus id
          type: integer
        - name: driver
          in: query
          required: true
          description: Driver id
          type: integer
        - name: route
          in: query
          required: true
          description: Route id
          type: integer
        - name: datetime
          in: query
          required: true
          description: Schedule
          type: string
        - name: isStandBy
          in: query
          required: true
          description: Bus stand by status
          type: boolean
    put:
      tags:
        - bus
      summary: Updates schedule for a bus
      description: Updates existing schedule record for a bus
      operationId: updateSchedule
      parameters:
        - name: id
          in: query
          required: true
          description: Schedule id
          type: integer
        - name: bus
          in: query
          required: true
          description: Bus id
          type: integer
        - name: driver
          in: query
          required: true
          description: Driver id
          type: integer
        - name: route
          in: query
          required: true
          description: Route id
          type: integer
        - name: datetime
          in: query
          required: true
          description: Schedule
          type: string
        - name: isStandBy
          in: query
          required: true
          description: Bus stand by status
          type: boolean
components:
  securitySchemes:
    AccessTokenAuth:
      type: apiKey
      in: header
      name: authorization
  responses:
    NotFound:
      description: The specified resource was not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    InvalidUser:
      description: Invalid username supplied
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    Error:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
      required:
        - code
        - message
    Bus:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
    Driver:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
    Route:
      type: object
      properties:
        id:
          type: integer
          format: int64
        checkpoints:
          type: array
          items:
            type: string
    Article:
      type: object
      properties:
        id:
          type: integer
          format: int64
        status:
          type: string
          enum:
            - draft
            - published
            - deleted
        title:
          type: string
          example: IG Won World Championship
        abstractContent:
          type: string
          example: Invictus Gaming sweeps Fnatic 3-0 to win League of Legends World Championship
        fullContent:
          type: string
          example: Article details goes here...
        sourceURL:
          type: string
          format: url
          example: http://www.espn.com/esports/story/_/id/25164094/invictus-gaming-sweeps-fnatic-3-0-win-league-legends-world-championship
        imageURL:
          type: string
          format: url
          example: https://a1.espncdn.com/combiner/i?img=%2Fphoto%2F2018%2F1103%2Fr458060_1296x729_16%2D9.jpg&w=920&h=518&scale=crop&cquality=80&location=origin&format=jpg
        timestamp:
          $ref: '#/components/schemas/Timestamp'
        platforms:
          type: array
          items:
            type: string
            enum:
              - a-platform
              - b-platform
              - c-platform
          uniqueItems: true
        disableComment:
          type: boolean
          example: false
        importance:
          type: integer
          format: int32
          enum:
            - 1
            - 2
            - 3
        author:
          type: string
          example: Syed Tajir Hasnain
        reviewer:
          type: string
          example: Syed Tajir Hasnain
        type:
          type: string
          enum:
            - CN
            - US
            - JP
            - EU
        pageviews:
          $ref: '#/components/schemas/Pageviews/properties/pageviews'
      required:
       - id
       - status
       - title
       - timestamp
       - importance
       - author
    Pageviews:
      type: object
      properties:
        key:
          type: string
          enum:
            - PC
            - Mobile
            - iOS
            - Android
        pageviews:
          type: integer
          format: int32
          example: 1024
      required:
        - key
        - pageviews
    Transaction:
      type: object
      properties:
        orderId:
          type: string
          format: uuid
        timestamp:
          $ref: '#/components/schemas/Timestamp'
        username:
          type: string
          example: Syed Tajir Hasnain
        price:
          type: number
          format: double
        status:
          type: string
          enum:
            - success
            - pending
      required:
        - orderId
        - timestamp
        - username
        - price
        - status
    Role:
      type: object
      properties:
        key:
          type: string
          example: admin
        name:
          type: string
          example: admin
        description:
          type: string
          example: Super Administrator. Have access to view all pages.
        routes:
          $ref: '#/components/schemas/AnyValue'
      required:
        - key
        - name
        - routes
    User:
      type: object
      properties:
        id:
          type: integer
          format: int64
        username:
          type: string
          example: Armour
        password:
          type: string
          example: 12345678
        name:
          type: string
          example: Syed Tajir Hasnain
        email:
          type: string
          example: armourcy@gmail.com
        phone:
          type: string
          example: 18888888888
        avatar:
          type: string
          format: url
        introduction:
          type: string
          example: Always Sunshine
        roles:
          type: array
          items:
            $ref: '#/components/schemas/Role/properties/key'
      required:
        - id
        - username
        - password
        - roles
    Timestamp:
      oneOf:
      - type: integer
        format: int32
        example: 1559605283
      - type: string
        format: date-time
        example: 2017-07-21T17:32:28Z
    AnyValue: {}